Fix Firefox_Boca's desktop entry
The default address used to access BOCA will change in the future. In
order to avoid problems, it was updated in the desktop entry.

A bug with random new windows appearing when opening links through
Firefox Boca has been fixed by defining a Window Manager Class Key in
the desktop entry.

As the standard Ubuntu 20.04 installation does not use Unity Shell, the
`OnlyShowIn` restriction to the configured actions has been removed.

- rebase file using Ubuntu's 20.04 default desktop entry for Firefox 87
- update boca's default address
- fix random new windows when opening links
- remove obsolete Unity Shell restriction
- add Window Manager Class (WMC) in a key and as a GTK launch option for
  Firefox