Update maintainer, dependencies and descriptions
The package's debian/control has been updated with the correct
information to contact the maintainer and links to the VCS where the
development happens.

The package maratona-desktop includes a .desktop file that references
firefox, which has to be added as an dependency. It was also added as a
dependency to maratona-desktop-latam.

The packages maratona-conflitos, maratona-task-data and
maratona-flatpak-common had their descriptions updated. This fixes the
Lintian warning extended-description-contains-empty-paragraph (see
issue #50 for more details).

- updates maintainer's name and email
- adds VCS information
- adds firefox as a dependency from maratona-desktop
- updates the descriptions of maratona-conflitos, maratona-task-data and
  maratona-flatpak-common (fixes Lintian warning
  extended-description-contains-empty-paragraph, check issue #50)
- adds maratona-desktop as a dependency from maratona-desktop-latam