Set exit on failure for postinst and postrm
The post removal and post installation scripts had the set -e flag added
in order to be conformant to the Debian Policy, aborting the execution
whenever commands, lists and compound bash commands throw uncaught
errors.

The modified scripts were debian/maratona-conflitos.postinst,
debian/maratona-conflitos.postrm, debian/maratona-desktop-latam.postrm,
debian/maratona-desktop-latam.preinst and
debian/maratona-usuario-icpc.postinst. This fixed multiple Lintian
warnings (maintainer-script-ignores-errors, see progress on issue #50).

The script maratona-usuario-icpc.postinst had to be modified to receive
the set -e flag, as it was indirectly checking the return status of a
command with `$?`. This approach would block the execution of the script
and could possibly create chalenges to the maintenability, as the usage
of `$?` to check the return status has many pitfalls.

- adds set -e for post installation and removal scripts in order to fix
  Lintian maintainer-script-ignores-errors, check issue #50)
- modify maratona-usuario-icpc.postinst to eliminate indirect status
  checks that would cause an abnormal execution