Fix the installation of VSCode extensions
Enable the full installation of the pre-installed extensions introduced
by the updated maratona-vscode-extensions on commit d44c9c3.

The Java configuration was removed, as the Java extension had
difficulties indentifying the system JDK. As IntelliJ studio is also a
flatpak and works as expected, there should be no negative impacts on
the user experience.