Enable specifying the package's version
Implement a function that enables the user to specify a version for the
package to be installed. This fixes the problems with non-reproducible
builds, as the Flatpaks were always downloaded in their newest versions
available on Flathub.

In order to specify a version, the user has to pass the OSTree commit as
an argument. Due to Flatpak's limitations, the newest package will be
installed first, and only then the specified version will replace it.

The new function uses the older one to install the standard version
offered by Flathub. This installation procedure, as expected from the
old behaviour, will be attempted three times.

If no version is specified, the new function will install the newest
version in the repository. If a version is specified, the function will
attempt to install it three times. If this installation fails, the
program will attempt to remove the package.

The OSTree commit id can be obtained with the command `flatpak
remote-info --log ${remote_name} ${package_name}`.

- create function for specifying a OSTree commit to be installed instead
  of the default one offered by the remote repository
- create function to remove a Flatpak package
- protect the arguments from the old function from unexpected expansions