Add conflict with snapd
Forcing a conflict with snapd will cause its removal when
maratona-conflitos is installed. The complete removal (purge) of snapd
and the installed snap packages will only happen if the administrator
passes the `--purge` flag to `apt install`.

This change will help with the minimisation of the image generated for
the contest.

- add conflict with snapd
- update description