Fix broken shortcuts and start-up experience
After the migration to Ubuntu 22.04 LTS, the shortcuts (.desktop files)
and the command used to suppress the Ubuntu's start-up screen stopped
working.

This commit refactors the `zera-home-icpc` in order to fix the
aforementioned problems. The GIO metadata for trusted .desktop files
was changed from *yes* to *true*, reflecting the upstream introduced
updates. A new dependency has been added to ensure that the DBUS will be
available in order to store the GIO metadata. The script has been also
rewritten to standardise the indentation, quoting and variable
expansion. Repeated commands were also replaced by Bash arrays.

- depend on dbus-x11
- update the values configured for the trusted metadata
- fix mixed indentation and repeated code
- correct shellcheck warnings